@@ -78,25 +78,25 @@
         "deploy:ci": "ng deploy --no-build --message=\"Release $npm_package_name (v$npm_package_version) on gh-pages\""
     },
     "dependencies": {
-        "@angular/animations": "^14.2.5",
-        "@angular/common": "^14.2.5",
-        "@angular/compiler": "^14.2.5",
-        "@angular/core": "^14.2.5",
-        "@angular/forms": "^14.2.5",
-        "@angular/localize": "^14.2.5",
-        "@angular/platform-browser": "^14.2.5",
-        "@angular/platform-browser-dynamic": "^14.2.5",
-        "@angular/platform-server": "^14.2.5",
-        "@angular/router": "^14.2.5",
-        "@ctrl/ngx-codemirror": "^5.1.1",
-        "@fortawesome/angular-fontawesome": "^0.11.1",
-        "@fortawesome/fontawesome-svg-core": "^6.2.0",
-        "@fortawesome/free-solid-svg-icons": "^6.2.0",
+        "@angular/animations": "^15.0.4",
+        "@angular/common": "^15.0.4",
+        "@angular/compiler": "^15.0.4",
+        "@angular/core": "^15.0.4",
+        "@angular/forms": "^15.0.4",
+        "@angular/localize": "^15.0.4",
+        "@angular/platform-browser": "^15.0.4",
+        "@angular/platform-browser-dynamic": "^15.0.4",
+        "@angular/platform-server": "^15.0.4",
+        "@angular/router": "^15.0.4",
+        "@codemirror/legacy-modes": "^6.3.1",
+        "@fortawesome/angular-fontawesome": "^0.12.0",
+        "@fortawesome/fontawesome-svg-core": "^6.2.1",
+        "@fortawesome/free-solid-svg-icons": "^6.2.1",
         "@kolkov/ngx-gallery": "2.0.1",
-        "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+        "@ng-bootstrap/ng-bootstrap": "^14.0.0",
         "@popperjs/core": "^2.11.6",
-        "bootstrap": "^5.2.2",
-        "codemirror": "^5.65.9",
+        "bootstrap": "^5.2.3",
+        "codemirror": "^6.0.1",
         "d3-drag": "^3.0.0",
         "d3-fetch": "^3.0.1",
         "d3-force": "^3.0.0",
@@ -104,54 +104,54 @@
         "d3-zoom": "^3.0.0",
         "font-awesome": "^4.7.0",
         "json2typescript": "^1.5.1",
-        "n3": "^1.16.2",
-        "ngx-json-viewer": "^3.1.0",
-        "rdfstore": "musicenfanthen/rdfstore-js#v0.9.18-alpha.6",
-        "rxjs": "~7.5.7",
+        "n3": "^1.16.3",
+        "ngx-json-viewer": "^3.2.1",
+        "rdfstore": "musicenfanthen/rdfstore-js#v0.9.18-alpha.7",
+        "rxjs": "~7.8.0",
         "snapsvg": "^0.5.1",
-        "tslib": "^2.4.0",
-        "zone.js": "~0.11.8"
+        "tslib": "^2.4.1",
+        "zone.js": "~0.12.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.2.5",
-        "@angular-eslint/builder": "^14.1.2",
-        "@angular-eslint/eslint-plugin": "^14.1.2",
-        "@angular-eslint/eslint-plugin-template": "^14.1.2",
-        "@angular-eslint/schematics": "^14.1.2",
-        "@angular-eslint/template-parser": "^14.1.2",
-        "@angular/cli": "^14.2.5",
-        "@angular/compiler-cli": "^14.2.5",
-        "@commitlint/cli": "^17.1.2",
-        "@commitlint/config-angular": "^17.1.0",
+        "@angular-devkit/build-angular": "^15.0.4",
+        "@angular-eslint/builder": "^15.1.0",
+        "@angular-eslint/eslint-plugin": "^15.1.0",
+        "@angular-eslint/eslint-plugin-template": "^15.1.0",
+        "@angular-eslint/schematics": "^15.1.0",
+        "@angular-eslint/template-parser": "^15.1.0",
+        "@angular/cli": "^15.0.4",
+        "@angular/compiler-cli": "^15.0.4",
+        "@commitlint/cli": "^17.3.0",
+        "@commitlint/config-angular": "^17.3.0",
         "@compodoc/compodoc": "^1.1.19",
         "@types/d3": "^7.4.0",
-        "@types/jasmine": "~4.3.0",
-        "@types/node": "^16.11.64",
-        "@typescript-eslint/eslint-plugin": "^5.39.0",
-        "@typescript-eslint/parser": "^5.39.0",
-        "angular-cli-ghpages": "^1.0.3",
+        "@types/jasmine": "~4.3.1",
+        "@types/node": "^16.18.10",
+        "@typescript-eslint/eslint-plugin": "^5.47.0",
+        "@typescript-eslint/parser": "^5.47.0",
+        "angular-cli-ghpages": "^1.0.5",
         "conventional-recommended-bump": "^6.1.0",
         "cross-var": "^1.1.0",
-        "eslint": "^8.25.0",
+        "eslint": "^8.30.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-angular": "^4.1.0",
         "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-jsdoc": "^39.3.6",
+        "eslint-plugin-jsdoc": "^39.6.4",
         "eslint-plugin-prettier": "^4.2.1",
-        "gzipper": "^7.1.0",
-        "husky": "^8.0.1",
-        "jasmine-core": "~4.4.0",
+        "gzipper": "^7.2.0",
+        "husky": "^8.0.2",
+        "jasmine-core": "~4.5.0",
         "karma": "~6.4.1",
         "karma-chrome-launcher": "~3.1.1",
         "karma-coverage": "^2.2.0",
         "karma-jasmine": "~5.1.0",
         "karma-jasmine-html-reporter": "^2.0.0",
-        "lint-staged": "^13.0.3",
-        "prettier": "^2.7.1",
+        "lint-staged": "^13.1.0",
+        "prettier": "^2.8.1",
         "source-map-explorer": "^2.5.3",
         "standard-version": "^9.5.0",
-        "typescript": "~4.7.4",
-        "webpack-bundle-analyzer": "^4.6.1"
+        "typescript": "~4.8.4",
+        "webpack-bundle-analyzer": "^4.7.0"
     },
     "engines": {
         "node": ">= 14.17.0",
