@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { DebugElement, SimpleChange } from '@angular/core';
-import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, ReactiveFormsModule } from '@angular/forms';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ReactiveFormsModule, UntypedFormBuilder, UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 
 import { FontAwesomeTestingModule } from '@fortawesome/angular-fontawesome/testing';
 import { faDiagramProject, faGripHorizontal, faTable } from '@fortawesome/free-solid-svg-icons';
@@ -9,8 +9,8 @@ import Spy = jasmine.Spy;
 
 import { expectSpyCall, getAndExpectDebugElementByCss } from '@testing/expect-helper';
 
-import { ViewHandle, ViewHandleTypes } from './view-handle.model';
 import { ViewHandleButtonGroupComponent } from './view-handle-button-group.component';
+import { ViewHandle, ViewHandleTypes } from './view-handle.model';
 
 describe('ViewHandleButtonGroupComponent (DONE)', () => {
     let component: ViewHandleButtonGroupComponent;
@@ -28,8 +28,8 @@ describe('ViewHandleButtonGroupComponent (DONE)', () => {
 
     beforeEach(async () => {
         await TestBed.configureTestingModule({
-            imports: [FontAwesomeTestingModule, ReactiveFormsModule],
-            declarations: [ViewHandleButtonGroupComponent, NgbTooltip],
+            imports: [FontAwesomeTestingModule, NgbTooltip, ReactiveFormsModule],
+            declarations: [ViewHandleButtonGroupComponent],
             providers: [UntypedFormBuilder],
         }).compileComponents();
     });
